code cleanup

+++ src/apps/src/haipa-zero/Configuration/Clients/SystemClientGenerator.cs
@@ -1,6 +1,5 @@
 ﻿using System;
 using System.IO;
-using Haipa.App;
 using Haipa.Configuration.Model;
 using Haipa.Security.Cryptography;
 using Newtonsoft.Json;
@@ -19,23 +18,19 @@ namespace Haipa.Runtime.Zero.Configuration.Clients
 
             AsymmetricKeyParameter publicKey = null;
             if (File.Exists(systemClientDataFile))
-            {
                 try
                 {
                     var systemClientData =
                         JsonConvert.DeserializeObject<ClientConfigModel>(File.ReadAllText(systemClientDataFile));
                     publicKey = CertHelper.GetPublicKey(systemClientData.X509CertificateBase64);
-
                 }
                 catch (Exception)
                 {
                     recreateSystemClient = true;
                 }
-            }
 
             AsymmetricCipherKeyPair privateKeyPair = null;
             if (!recreateSystemClient && File.Exists(systemClientKeyFile))
-            {
                 try
                 {
                     privateKeyPair = CertHelper.ReadPrivateKeyFile(systemClientKeyFile);
@@ -44,13 +39,10 @@ namespace Haipa.Runtime.Zero.Configuration.Clients
                 {
                     recreateSystemClient = true;
                 }
-            }
 
             if (!recreateSystemClient && publicKey != null && privateKeyPair != null)
-            {
                 if (privateKeyPair.Public.Equals(publicKey))
                     return;
-            }
 
             RemoveSystemClient();
 
@@ -60,13 +52,12 @@ namespace Haipa.Runtime.Zero.Configuration.Clients
             {
                 ClientId = "system-client",
                 X509CertificateBase64 = Convert.ToBase64String(certificate.GetEncoded()),
-                AllowedScopes = new []{ "openid", "compute_api", "common_api", "identity:clients:write:all" }
+                AllowedScopes = new[] {"openid", "compute_api", "common_api", "identity:clients:write:all"}
             };
 
             var clientIO = new ConfigIO(ZeroConfig.GetClientConfigPath());
             clientIO.SaveConfigFile(newClient, newClient.ClientId);
             CertHelper.WritePrivateKeyFile(systemClientKeyFile, keyPair);
-
         }
 
         private static void RemoveSystemClient()
@@ -74,16 +65,9 @@ namespace Haipa.Runtime.Zero.Configuration.Clients
             var systemClientDataFile = Path.Combine(ZeroConfig.GetClientConfigPath(), "system-client.json");
             var systemClientKeyFile = Path.Combine(ZeroConfig.GetClientConfigPath(), "system-client.key");
 
-            if (File.Exists(systemClientDataFile))
-            {
-                File.Delete(systemClientDataFile);
-            }
+            if (File.Exists(systemClientDataFile)) File.Delete(systemClientDataFile);
 
-            if (File.Exists(systemClientKeyFile))
-            {
-                File.Delete(systemClientKeyFile);
-            }
+            if (File.Exists(systemClientKeyFile)) File.Delete(systemClientKeyFile);
         }
-
     }
 }
